requirements.txt
amqp==5.0.9
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
debugpy==1.5.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-spectacular==0.21.2
flower==1.0.0
gunicorn==20.1.0
humanize==4.0.0
inflection==0.5.1
jsonschema==4.4.0
kombu==5.2.3
packaging==21.3
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
wrapt==1.13.3
django-jazzmin==2.4.9